Move Conn and Socket to a separate module
